Robot follows wall, more or less
This is the first attempt to have the robot actually walk parallel to a
wall.  It sort of works.  The robot can adjust its heading based on
feedback from the side ultrasonic sensors, and does stop when an
obstacle is in front.  The steering seems to be swamped by occasional
bad data.  For instance the wall will usually be say 20-20, 20-20, 19-20
cm from the sensors but every 8th reading might be an obvious error of
3mm.  I'm trying to see if the bad data is only present when the motors
are running.  A little averaging might help clean the data, as well as
throwing out readings which are sudden changes.  Also the NewPing
library apparently has good averaging built in.